+++ src/app/Components/Charts/Speedometer/MaxRiskPercent.tsx
@@ -15,7 +15,6 @@ const MaxRiskSpeedometer = ({ metric, min, max, colorArray, labelArray, title }:
     return (
         <div className="boxData" style={{
             height: '250px',
-            margin: '25px',
             minWidth: '300px',
             maxWidth: '300px'
         }}>