+++ src/app/Pages/ActiveDeals/StyledDiv.tsx
@@ -8,10 +8,12 @@ export default styled.div`
     border-spacing: 0;
     background-color: var(--color-background-light);
     color: var(--color-text-lightbackground);
-    font-size: .875em;
+    font-size: .9em;
     min-width: 1200px;
     width: 100%;
     
+    .test {
+    }
 
     th,
     td {
@@ -23,13 +25,16 @@ export default styled.div`
         position: sticky;
         top: 0;
         z-index: 100;
+        background-color: orange;
+        padding: 0;
     }
 
     thead tr:nth-child(2) th {
         position: sticky;
-        top: 53px;
+        top: 44px;
         z-index: 100;
     }
+
 
     
     tbody{
@@ -47,12 +52,15 @@ export default styled.div`
         };
 
         .pill {
-            padding: .1em;
+            padding: .2em;
             display: block;
             border-radius: 10px;
             color: white;
-            font-weight: 700;
+            font-weight: 600;
+            white-space: nowrap;
         }
+
+
 
         .red {
             background-color: var(--color-red);
@@ -61,7 +69,21 @@ export default styled.div`
 
         .green {
             background-color: var(--color-green);
-            opacity: var(--opacity-pill)
+            opacity: var(--opacity-pill);
+        }
+
+        .pill-left {
+            border-radius: 10px 0px 0px 10px;
+            text-align: right;
+            padding-right: 1em;
+            margin-right: 2px;
+        }
+
+        .pill-right {
+            border-radius: 0px 10px 10px 0px;
+            text-align: left;
+            padding-left: 1em;
+            height: 100%;
         }
     }